requirements.txt
Pillow==10.1.0
boto3==1.33.6
requests==2.31.0
tqdm==4.66.1
PyYAML==6.0.1
